+++ eNMS/services/notification/mattermost_notification.py
@@ -22,11 +22,11 @@ class MattermostNotificationService(Service):
     __mapper_args__ = {"polymorphic_identity": "mattermost_notification_service"}
 
     def job(self, run, payload, device=None):
-        channel = run.sub(run.channel, locals()) or app.config["mattermost"]["channel"]
+        channel = run.sub(run.channel, locals()) or app.settings["mattermost"]["channel"]
         run.log("info", f"Sending MATTERMOST notification on {channel}", device)
         result = post(
-            app.config["mattermost"]["url"],
-            verify=app.config["mattermost"]["verify_certificate"],
+            app.settings["mattermost"]["url"],
+            verify=app.settings["mattermost"]["verify_certificate"],
             data=dumps({"channel": channel, "text": run.sub(run.body, locals())}),
         )
         return {"success": True, "result": str(result)}